前台查询+发起政策申请+后台bug修复+后台人数统计

+++ talentweb/src/main/java/com/talentcard/web/vo/PolicyDetailVO.java
@@ -172,7 +172,7 @@ public class PolicyDetailVO implements Serializable {
 
     private String responsibleUnit;
     //符合条件人数
-    private Integer meetConditionNumber;
+    private Long meetConditionNumber;
     /**
      * po 转 vo
      *
